feat: util de distancias das lojas
